# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "岚山区劳动争议调解区块链平台，后台采用golang开发，使用gin + gorm + gorbac + jwt开发",
        "title": "岚山区劳动争议调解区块链平台",
        "termsOfService": "http://hafrans.com",
        "contact": {
            "name": "Chuuka Ro (Hafrans)",
            "url": "http://hafrans.com/support",
            "email": "lvzh@hafrans.com"
        },
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/case/": {
            "get": {
                "description": "获取所有的案件，非管理员只能看到自己的，管理员能看到全部人的",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labor"
                ],
                "summary": "获取所有的案件",
                "parameters": [
                    {
                        "type": "number",
                        "description": "页码",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "number",
                        "description": "页大小",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "案件号（case id）模糊查询",
                        "name": "search",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常业务处理",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/case/caseId/:caseId": {
            "get": {
                "description": "获取单一Case，通过CaseID 例子：3711002020063019254015935163407436142",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case"
                ],
                "summary": "通过Case ID（调解申请号）获取case",
                "parameters": [
                    {
                        "type": "string",
                        "description": "不是案件号码,不是案件id",
                        "name": "caseId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/case/create": {
            "post": {
                "description": "由申请人填写创建新调解案件",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case"
                ],
                "summary": "创建新调解案件",
                "parameters": [
                    {
                        "description": "提交表单",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.CaseFirstSubmitForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/case/delete/:id": {
            "get": {
                "description": "删除单一Case，一般用户只可以删除自己的，特殊权限者可以删除任何人的，注意：如果案件正在处理中，则无法删除",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case"
                ],
                "summary": "通过 ID（调解案件ID，不是case_id）删除case",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "案件id,不是案件号码",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/case/id/:id": {
            "get": {
                "description": "获取单一Case， 例子：9",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case"
                ],
                "summary": "通过ID（主键）获取case",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "案件id,不是案件号码",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/case/status_change/:id": {
            "post": {
                "description": "申请人只可以确认、拒绝状态；管理人员可以设置任何状态 StatusSubmitted= 0 已提交；StatusPending = 1 正在处理；StatusResultConfirming = 2 当事人等待确认调解结果；StatusRefused =3 拒绝调解；StatusConfirmed=4 确认调解；StatusCompleted=5；结束调解               // 调解结束",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case"
                ],
                "summary": "修改调解案件的状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "案件模型id，注意：不是案件号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "提交表单",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.CaseStatusChangeForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/case/update/:id": {
            "post": {
                "description": "由申请人填写修改调解案件，申请人、管理员、部门人员可以修改，只可以修改案件信息，不可以修改案件状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case"
                ],
                "summary": "修改调解案件",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "案件模型id，注意：不是案件号",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "提交表单",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.CaseFirstSubmitForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/category": {
            "get": {
                "description": "获取所有类型",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "category",
                    "case"
                ],
                "summary": "获得所有类型",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
        "/api/v1/test/case/template": {
            "get": {
                "description": "获取申请调解案件的上传模板，测试用",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case",
                    "test"
                ],
                "summary": "获取申请调解案件的上传模板",
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "vo.Applicant": {
            "type": "object",
            "required": [
                "applicant_address",
                "applicant_birth",
                "applicant_contact",
                "applicant_id",
                "applicant_name",
                "applicant_nationality"
            ],
            "properties": {
                "applicant_address": {
                    "description": "地址",
                    "type": "string"
                },
                "applicant_birth": {
                    "description": "生日",
                    "type": "string"
                },
                "applicant_contact": {
                    "description": "联系方式",
                    "type": "string"
                },
                "applicant_id": {
                    "description": "身份证号",
                    "type": "string"
                },
                "applicant_name": {
                    "description": "姓名",
                    "type": "string"
                },
                "applicant_nationality": {
                    "description": "民族",
                    "type": "string"
                }
            }
        },
        "vo.CaseFirstSubmitForm": {
            "type": "object",
            "required": [
                "applicant",
                "category_id",
                "content",
                "form_id",
                "respondent",
                "title"
            ],
            "properties": {
                "applicant": {
                    "description": "1.申请人",
                    "type": "object",
                    "$ref": "#/definitions/vo.Applicant"
                },
                "category_id": {
                    "description": "类型id",
                    "type": "integer"
                },
                "content": {
                    "description": "4.调解事项所依据的事实与理由",
                    "type": "string"
                },
                "form_id": {
                    "description": "5.劳动者填写的表单 id",
                    "type": "integer"
                },
                "materials": {
                    "description": "6. 证据材料",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/vo.Material"
                    }
                },
                "respondent": {
                    "description": "2.被申请人",
                    "type": "object",
                    "$ref": "#/definitions/vo.Employer"
                },
                "title": {
                    "description": "3.调解事项",
                    "type": "string"
                }
            }
        },
        "vo.CaseStatusChangeForm": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "vo.Common": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
        "vo.CommonData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
...
        "vo.Employer": {
            "type": "object",
            "required": [
                "employer_address",
                "employer_contact",
                "employer_faren",
                "employer_name",
                "employer_uscc"
            ],
            "properties": {
                "employer_address": {
                    "description": "地址",
                    "type": "string"
                },
                "employer_contact": {
                    "description": "联系方式",
                    "type": "string"
                },
                "employer_faren": {
                    "description": "法人",
                    "type": "string"
                },
                "employer_name": {
                    "description": "公司名",
                    "type": "string"
                },
                "employer_uscc": {
                    "description": "识别号",
                    "type": "string"
                }
            }
        },
...
        "vo.Material": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "材料介绍",
                    "type": "string"
                },
                "path": {
                    "description": "材料资源path",
                    "type": "string"
                }
            }
        },
...
    }
}